third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/cross-origin-opener-policy/reporting/access-reporting/access-to-coop-page-from-openee_coop-ro.https-expected.txt

This is a testharness.js-based test.
[FAIL] access-to-coop-page-from-openee, same-origin
  assert_equals: expected (string) "coop" but got (undefined) undefined
[FAIL] access-to-coop-page-from-openee, same-origin + redirect
  assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
